src/backend/pyproject.toml
[project]
name = "rag_app"
version = "1.0.0"
description = "Create a RAG application with PostgreSQL"
dependencies = [
    "fastapi>=0.111.0,<1.0.0",
    "uvicorn>=0.30.1,<1.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "environs>=11.0.0,<12.0.0",
    "aiohttp>=3.9.5,<4.0.0",
    "asyncpg>=0.29.0,<1.0.0",
    "SQLAlchemy[asyncio]>=2.0.30,<3.0.0",
    "pgvector>=0.2.5,<0.3.0",
    "openai>=1.34.0,<2.0.0",
    "tiktoken>=0.7.0,<0.8.0",
    "openai-messages-token-helper>=0.1.8,<0.2.0"
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"
